Migrate to `eslint.config.js` (#15748)
