+++ eslint/babel-eslint-config-internal/package.json
@@ -9,6 +9,10 @@
     ".": "./index.js",
     "./package.json": "./package.json"
   },
+  "dependencies": {
+    "@eslint/js": "^8.44.0",
+    "globals": "^13.20.0"
+  },
   "peerDependencies": {
     "@babel/eslint-parser": "^7.16.0"
   }
